Split nomos-libp2p crate into multiple modules
